Use `Fiber.defer` in standalone setup
and extract common code into ext/setup
